+++ src/prost/interchain_security.ccv.consumer.v1.rs
@@ -416,7 +416,8 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_next_fee_distribution(request).await
+                                <T as Query>::query_next_fee_distribution(&inner, request)
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -460,7 +461,7 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_params(request).await
+                                <T as Query>::query_params(&inner, request).await
                             };
                             Box::pin(fut)
                         }
@@ -506,7 +507,7 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_provider_info(request).await
+                                <T as Query>::query_provider_info(&inner, request).await
                             };
                             Box::pin(fut)
                         }