+++ src/prost/interchain_security.ccv.provider.v1.rs
@@ -360,7 +360,7 @@ pub mod msg_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).assign_consumer_key(request).await
+                                <T as Msg>::assign_consumer_key(&inner, request).await
                             };
                             Box::pin(fut)
                         }
@@ -406,7 +406,8 @@ pub mod msg_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).submit_consumer_misbehaviour(request).await
+                                <T as Msg>::submit_consumer_misbehaviour(&inner, request)
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -452,7 +453,8 @@ pub mod msg_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).submit_consumer_double_voting(request).await
+                                <T as Msg>::submit_consumer_double_voting(&inner, request)
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -1602,7 +1604,7 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_consumer_genesis(request).await
+                                <T as Query>::query_consumer_genesis(&inner, request).await
                             };
                             Box::pin(fut)
                         }
@@ -1648,7 +1650,7 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_consumer_chains(request).await
+                                <T as Query>::query_consumer_chains(&inner, request).await
                             };
                             Box::pin(fut)
                         }
@@ -1697,7 +1699,8 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_consumer_chain_starts(request).await
+                                <T as Query>::query_consumer_chain_starts(&inner, request)
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -1746,7 +1749,8 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_consumer_chain_stops(request).await
+                                <T as Query>::query_consumer_chain_stops(&inner, request)
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -1795,7 +1799,8 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_validator_consumer_addr(request).await
+                                <T as Query>::query_validator_consumer_addr(&inner, request)
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -1844,7 +1849,8 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_validator_provider_addr(request).await
+                                <T as Query>::query_validator_provider_addr(&inner, request)
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -1890,7 +1896,7 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_throttle_state(request).await
+                                <T as Query>::query_throttle_state(&inner, request).await
                             };
                             Box::pin(fut)
                         }
@@ -1939,7 +1945,11 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner).query_throttled_consumer_packet_data(request).await
+                                <T as Query>::query_throttled_consumer_packet_data(
+                                        &inner,
+                                        request,
+                                    )
+                                    .await
                             };
                             Box::pin(fut)
                         }
@@ -1988,8 +1998,10 @@ pub mod query_server {
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
-                                (*inner)
-                                    .query_registered_consumer_reward_denoms(request)
+                                <T as Query>::query_registered_consumer_reward_denoms(
+                                        &inner,
+                                        request,
+                                    )
                                     .await
                             };
                             Box::pin(fut)